WIP solve most of reactos gc warnings

+++ drivers/filesystems/udfs/udf_info/extent.cpp
@@ -2271,7 +2271,7 @@ UDFResizeExtent(
 
     UDFCheckSpaceAllocation(Vcb, 0, ExtInfo->Mapping, AS_USED); // check if used
     if(ExtInfo->Offset) {
-        if(ExtInfo->Offset + Length <= LBS) {
+        if(ExtInfo->Offset + Length <= (int64)LBS) {
             ExtPrint(("Resize IN-ICB\n"));
             ExtInfo->Length = Length;
             return STATUS_SUCCESS;
@@ -2501,7 +2501,7 @@ UDFResizeExtent(
             if(l < Length) {
                 // we get here if simple increasing of the last frag failed
                 AdPrint(("Resize add new frag (7)\n"));
-                if(l < LBS && Length >= LBS &&
+                if(l < (int64)LBS && Length >= (int64)LBS &&
                    (ExtInfo->Flags & EXTENT_FLAG_ALLOC_MASK) == EXTENT_FLAG_ALLOC_SEQUENTIAL) {
                     AdPrint(("Resize tune for SEQUENTIAL i/o\n"));
                 }
@@ -2987,7 +2987,7 @@ UDFReadExtentCached(
     OSSTATUS status;
     // prevent reading out of data space
     if(Offset > ExtInfo->Length) return STATUS_END_OF_FILE;
-    if(Offset+Length > ExtInfo->Length) Length = (uint32)(ExtInfo->Length - Offset);
+    if(Offset+Length > (uint64)ExtInfo->Length) Length = (uint32)(ExtInfo->Length - Offset);
     Offset += ExtInfo->Offset;               // used for in-ICB data
     // read maximal possible part of each frag of extent
     while(((LONG)Length) > 0) {
@@ -3041,7 +3041,7 @@ UDFReadExtent(
     OSSTATUS status;
     // prevent reading out of data space
     if(Offset > ExtInfo->Length) return STATUS_END_OF_FILE;
-    if(Offset+Length > ExtInfo->Length) Length = (uint32)(ExtInfo->Length - Offset);
+    if(Offset+Length > (uint64)ExtInfo->Length) Length = (uint32)(ExtInfo->Length - Offset);
     Offset += ExtInfo->Offset;               // used for in-ICB data
     // read maximal possible part of each frag of extent
     Lba = UDFExtentOffsetToLba(Vcb, Extent, Offset, &sect_offs, &to_read, &flags, &index);
